+++ apps/core/cmd/supervisor/router/register.go
@@ -1,13 +1,9 @@
 package router
 
 import (
-	"context"
 	"fmt"
-	"log"
 
 	"nexus/internal/protocol"
-	superdb "nexus/internal/supervisor/db/gen"
-	"nexus/internal/supervisor/store"
 	"nexus/internal/transport"
 	"nexus/internal/util"
 )
@@ -35,21 +31,21 @@ func (r *supervisorRouter) register(conn *transport.Conn, body *protocol.Registe
 		}
 		r.workers.Append(key, conn)
 
-		q := store.GetStorePool().Queries()
-		ctx := context.Background()
-		n, err := q.GetNode(ctx, superdb.GetNodeParams{ID: 3, OwnerUserID: 3})
-		u, err := q.GetUser(ctx, n.OwnerUserID)
-		if err == nil {
+		// q := store.GetStorePool().Queries()
+		// ctx := context.Background()
+		// n, err := q.GetNode(ctx, superdb.GetNodeParams{ID: 3, OwnerUserID: 3})
+		// u, err := q.GetUser(ctx, n.OwnerUserID)
+		// if err == nil {
 
-			err := r.Exec(u, key, protocol.ExecTypeExec, n)
-			if err != nil {
-				log.Printf("[PROCESS ERR] %v", err)
-			} else {
-				log.Printf("[PROCESS SUC]")
-			}
-		} else {
-			log.Printf("[ERR] %v", err)
-		}
+		// 	err := r.Exec(u, key, protocol.ExecTypeExec, n)
+		// 	if err != nil {
+		// 		log.Printf("[PROCESS ERR] %v", err)
+		// 	} else {
+		// 		log.Printf("[PROCESS SUC]")
+		// 	}
+		// } else {
+		// 	log.Printf("[ERR] %v", err)
+		// }
 
 		return protocol.RegisterResponse{SubKey: body.SubKey}, nil
 	}